+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2025-04-07
+### Added
+- Add reset function
+- Organize code for better Readability and Maintainability 
+- Add new detailed results
+- Add input validation and checking
+- Add few JSDoc
+
 ## [0.1.0] - 2025-04-03
 ### Added
 - Initial project setup